Temporarily skip qtactiveqt testing on platforms other than Windows
ActiveQt currently gives a fatal error if you attempt to compile it on
any platform other than Windows.  Therefore the test script needs a
special hack in place to recognize that ActiveQt should not be compiled
on most platforms.

There are pending changes which will allow this hack to be removed,
but the changes affect both qt5 and qtactiveqt.  CI will be deadlocked
without this temporary hack.