tests: skip instead of failing if QT_MODULE_TO_TEST is unset
The qtqa module is a special case with respect to autotests.
It contains some autotests which are intended to be run once per module
to test those modules.  It also may contain autotests for testing qtqa
itself (just like all the other modules).

Therefore the expected test procedure is something like:

  qtqa: make -j1 -k check
  (... later)
  qtqa: QT_MODULE_TO_TEST=$PWD/../qtbase        make -j1 -k check
  qtqa: QT_MODULE_TO_TEST=$PWD/../qtdeclarative make -j1 -k check
  (... etc)

Currently, this breaks at the first step, as some tests will immediately
fail if QT_MODULE_TO_TEST is not set.  Make it skip instead, so that
`make check' just works for qtqa, as it is intended to for all modules.

Task: QTQAINFRA-145
Reviewed-by: Jason McDonald
Change-Id: I69384adad68db9985d2bb705a253a4f619797816