testrunner: fixed timeout stuck at 60 seconds
If the --timeout option was passed with a value higher than 60,
it would behave the same as `--timeout 60'.
This occurred because we were setting Proc::Reliable's maxtime()
(the maximum time for all tries), but forgetting to set
Proc::Reliable's time_per_try() (the maximum time for each try).

We don't use Proc::Reliable's retry feature at all, so set both timeouts
to the same value.

Reviewed-by: Kalle Lehtonen
Change-Id: I42ac2fcee9fac8732933914633527d4c814ccf93