setup.pl: do not print error-like messages unless there's really an error
setup.pl tests for the existence of perl and python modules.

Previously, it was printing various error messages about module failures
even in the successful case.  Adjust it so that these messages are only
printed in the failure case.  Otherwise, the errors confuse humans and
machines attempting to parse the test logs.

Task: QTQAINFRA-163
Reviewed-by: Kalle Lehtonen
Change-Id: If482a0352ef1cba291f08c4a5054b74347f507e2