testrunner: add --capture-logs, --tee-logs feature
This feature allows test logs for all tests to be safely captured into
a specified directory.  For example:

  $ make -j1 check "TESTRUNNER=testrunner --tee-logs=$HOME/test-logs --"

...from the top-level of the build tree will capture one log file per
autotest into $HOME/test-logs, as well as printing the output to the
screen as usual.

The intent of this feature is to make it safe and easy to capture test
logs for use as artifacts in continuous integration systems such as
Pulse.

Reviewed-by: Kalle Lehtonen
Reviewed-by: Keith Isdale
Change-Id: I1d93c4dccf05c0931893c46cc84a5cb1a294765a