testrunner: fixed 10-testrunner.t when using piping core pattern
If a core_pattern is used which pipes the core dump to a program, such
as abrt on Fedora, it seems that the RLIMIT_CORE rlimit is ignored.
This means there is no guaranteed way for this autotest to disable core
dumps, so it must be modified to handle both cases where core may or
may not be dumped.

Change-Id: Idc4e03610d825d54d0eb8cd2d52ca56c4d454bea
Reviewed-on: http://codereview.qt.nokia.com/649
Reviewed-by: Qt Sanity Bot <qt_sanity_bot@ovi.com>
Reviewed-by: Sergio Ahumada <sergio.ahumada@nokia.com>