qtmod_test: update nested gitmodules whenever changing a module's SHA1
The qtdeclarative gitmodule itself now contains a reference to another
gitmodule.  Normally, this is set up correctly by the init-repository
script; however, we are changing the SHA1 of the qtdeclarative
module _after_ we run init-repository, which means the state of the
nested gitmodule becomes out of sync.

Fix it by always doing an extra submodule update whenever we change the
SHA1 of any gitmodule.

Change-Id: I4252a8aae7885c5df5a86ddb328b17b55eb419bf
Reviewed-on: http://codereview.qt.nokia.com/860
Reviewed-by: Qt Sanity Bot <qt_sanity_bot@ovi.com>
Reviewed-by: Kalle Lehtonen
Reviewed-by: Rohan McGovern <rohan.mcgovern@nokia.com>