qtmod_test: use `qt-modularization-base' branch for qtwebkit
qtwebkit is unlike the rest of the Qt modules because we do not want
to track the `master' branch.  Previously, we used whatever qtwebkit
SHA1 was set in qt/qt5.git, and this happened to match the HEAD of
qt-modularization-base; now there is a build fix in the
qt-modularization-base branch which we want to make use of for all
modules, without waiting for it to get into qt/qt5.git.

Change-Id: I5344592fe5a272fe5e09a4929a8881c9bbc444ec
Reviewed-on: http://codereview.qt.nokia.com/967
Reviewed-by: Qt Sanity Bot <qt_sanity_bot@ovi.com>
Reviewed-by: Kalle Lehtonen
Reviewed-by: Rohan McGovern <rohan.mcgovern@nokia.com>