testrunner: add a `flaky' plugin to deal with unstable tests
With the `flaky' plugin activated, any failing test will be re-run
in an attempt to ascertain if the test failure is stable.

If the test does not fail on the second attempt, or it fails for a
different reason, the plugin will loudly complain that the autotest
appears to be unstable.

Change-Id: Ie9712c2448205cfa87b72f718e2cd21c102b0d3c
Reviewed-on: http://codereview.qt.nokia.com/2129
Reviewed-by: Qt Sanity Bot <qt_sanity_bot@ovi.com>
Reviewed-by: Sergio Ahumada <sergio.ahumada@nokia.com>
Reviewed-by: Rohan McGovern <rohan.mcgovern@nokia.com>