Automatically recover from some transient errors in test scripts
When running complex test procedures, it is unfortunately relatively
common that certain types of errors may arise which are unrelated to the
code under test.

One example of this is a test procedure which attempts to perform a git
clone from a remote host.  If implemented simply, this test procedure
will fail whenever the remote host is undergoing maintenance or a
temporary network outage has occurred.  This is usually undesirable.

This commit adds some infrastructure for test scripts to transparently
recover from these types of errors without having to explicitly code
for them.

Currently handled are various errors from git, ssh and scp relating
to network problems.

Task: QTQAINFRA-223
Change-Id: I8fb0a505cdbb8ad00f6bd036854b30e24201419b
Reviewed-on: http://codereview.qt.nokia.com/644
Reviewed-by: Qt Sanity Bot <qt_sanity_bot@ovi.com>
Reviewed-by: Kalle Lehtonen <kalle.ju.lehtonen@nokia.com>
Reviewed-by: Sergio Ahumada <sergio.ahumada@nokia.com>