qtmod_test: add an option to consider all tests as "insignificant"
If qt.tests.insignificant is 1, failures are discarded from all tests
(as if they were all marked with CONFIG+=insignificant_test).

The primary use case for this option is to ease the addition of new test
configurations.  For a new configuration, autotests may be first enabled
in non-enforcing mode, then later enabled in strict mode after handling
any observed issues.

Change-Id: Iab741ca8c148f5c750960b021f7bdf9d86d9ee7a
Reviewed-on: http://codereview.qt.nokia.com/2453
Reviewed-by: Kalle Lehtonen <kalle.ju.lehtonen@nokia.com>