qt_update_submodules: post a summary of the updates into gerrit
Currently the updates are very difficult to review, since the raw
SHA1s are the only changes displayed.  The reviewer has no choice but to
either check out the change locally and peruse the logs, or to not
really review the changes at all.

Let the script help by posting the content of "git submodule summary"
into gerrit for the reviewer to look at.

Change-Id: Id9d6e4c2619e846a9d1a75e7073813cea31d6f2e
Reviewed-on: http://codereview.qt.nokia.com/2888
Reviewed-by: Qt Sanity Bot <qt_sanity_bot@ovi.com>
Reviewed-by: Sergio Ahumada <sergio.ahumada@nokia.com>