test: skip one failing test on mac
This test assumes that the current process doesn't have the rights to
create files under / (as long as it isn't running as root).
This worked fine, but recently on some of our macs the permissions on /
were mysteriously changed for reasons which aren't yet understood.

For now, play it safe and skip anything which might write to / on mac.

Change-Id: I5f54b8e83c5c5895e9592bc4ceabacca13d18008
Reviewed-on: http://codereview.qt.nokia.com/3243
Reviewed-by: Qt Sanity Bot <qt_sanity_bot@ovi.com>
Reviewed-by: Rohan McGovern <rohan.mcgovern@nokia.com>