test: replace QT_MODULE_KERNEL with QLibraryInfo to find Qt lib dir
QLibraryInfo is the canonical way to programmatically obtain the
location of the Qt libs dir.  Previously this test used an environment
variable named QT_MODULE_KERNEL.  No test script has ever contained the
logic for setting this variable.

Also, using QLibraryInfo avoids assuming that the libs directory is
located at $PREFIX/lib, which is not necessarily the case for custom
builds.

Change-Id: I41d8c8cc1647969f022d22bc5a47d862ebf396b3
Reviewed-on: http://codereview.qt.nokia.com/3283
Reviewed-by: Qt Sanity Bot <qt_sanity_bot@ovi.com>
Reviewed-by: Keith Isdale <keith.isdale@nokia.com>