qtmod_test: support testing of modules hosted outside of qt/qt5.git
Usually we are using the qt.pro in qt5.git as the canonical source of
dependency information.  This means the script wasn't able to test any
modules hosted outside of qt5.git.

Add basic support for testing modules outside of qt5.git by using the
dependency information in sync.profile.  This currently works only for
a single level of dependencies (i.e. it will not work for qtaddon1
depends on qtaddon2 depends on qtbase, where qtaddon1 and qtaddon2
are both hosted outside of qt5.git).

Change-Id: Ia8b04174d43a9b34fad5242e040b3995116cf372
Reviewed-on: http://codereview.qt.nokia.com/3534
Reviewed-by: Qt Sanity Bot <qt_sanity_bot@ovi.com>
Reviewed-by: Sergio Ahumada <sergio.ahumada@nokia.com>