Create license header autotest
This script replaces the Qt4 license checker from the qa-dungeon
repository and has knowledge of the module structure of Qt5.

The test will scan source files in the directory nominated by the
$QT_MODULE_TO_TEST enviroment variable and verifies whether or not each
source file that must contain a valid Qt license header actually does.
The reference license text for each license type is obtained from the
header.* files in the qtbase module, which is expected to be present in
$QT_MODULE_TO_TEST/../qtbase.

Change-Id: Idcff8774ba57ec31981adc5ea6b95fdb50b83531
Reviewed-on: http://codereview.qt-project.org/4511
Reviewed-by: Qt Sanity Bot <qt_sanity_bot@ovi.com>
Reviewed-by: Rohan McGovern <rohan.mcgovern@nokia.com>