Move qtqa shared tests to tests/shared.
Having these tests under tests/auto is misleading: usually this
directory contains tests for the containing module, but these tests are
intended to be shared tests for _all_ Qt modules.

Move them to tests/shared instead to make this difference clearer.

Note the tests are now compiled by default, but not run by default.
To run these tests, it's necessary to explicitly opt-in by doing a
`make check' under the tests/shared directory.

Change-Id: I42559710de8e70b708612da45c366d8d00e16ad6
Reviewed-on: http://codereview.qt-project.org/4855
Reviewed-by: Qt Sanity Bot <qt_sanity_bot@ovi.com>
Reviewed-by: Jason McDonald <jason.mcdonald@nokia.com>