parse_build_log: fixed `spam' from errors relating to common filenames
Previously, if we had found that some source file failed to compile, we
would extract _all_ lines from the log which referred to that file.

This causes a lot of useless messages when errors relate to a source
file with a common name, such as main.cpp (there are more than 400
such files in qtbase alone).

Use smarter logic: when we hit a compile fail error, only extract lines
referring to the source file from nearby in the log, instead of the
whole log.

As a side benefit, this removes the clumsy compile_fail_source_re and
compile_fail_exclude mechanisms.

Change-Id: I23c9465c937790e4c67490f579385b829d1fc0be
Reviewed-on: http://codereview.qt-project.org/5035
Reviewed-by: Kalle Lehtonen <kalle.ju.lehtonen@nokia.com>