testrunner: make capturing work with testlib multi-destination logging
testlib recently gained the ability to write more than one log file at a
time, or to write both to standard output and a log file.
Our testrunner script is designed to capture test logs as they are
written, so it now needs to be updated to capture multiple logs
simultaneously.

Note that this somewhat obsoletes the `--tee-logs' feature of this
script, since testlib itself now has tee-like behavior.

This is an enabler for switching our CI to get the best of both worlds
in the test reports: we can print human-friendly plain text straight
to the build log, and machine-friendly testlib XML to a file on disk.

Change-Id: I185925955cc5372b4a54c284c5f99637a40a67aa
Reviewed-on: http://codereview.qt-project.org/6210
Sanity-Review: Qt Sanity Bot <qt_sanity_bot@ovi.com>
Reviewed-by: Jason McDonald <jason.mcdonald@nokia.com>