Prefer QSKIP to QTEST_NOOP_MAIN.
These tests depend on QProcess, but there is no way for a .pro file to
know if QProcess is in the Qt build, so the tests cannot be omitted from
the build if they are inapplicable.  Because of this limitation, the
tests were using QTEST_NOOP_MAIN when QProcess was not available, making
the tests appear to pass.  This commit changes QTEST_NOOP_MAIN to QSKIP,
so that the user receives a clear indication that the tests aren't
testing anything when QProcess is not available.

Change-Id: I2ea5ccd6062146e888bc91ad05a929ff8c532bee
Reviewed-by: Rohan McGovern <rohan.mcgovern@nokia.com>