tst_headers: remove the qobjectdefs.h special case
This header was special cased for the Q_PRIVATE_SLOTS check because this
is the header which contains `#define Q_PRIVATE_SLOTS'.

Change the test to explicitly skip only that line instead, so that
the rest of the file is checked and the test will not start failing if
the file is renamed.

Change-Id: Id6af0a480efc9fce48650c734ea10baf8d7515eb
Sanity-Review: Qt Sanity Bot <qt_sanity_bot@ovi.com>
Reviewed-by: Jason McDonald <jason.mcdonald@nokia.com>
Reviewed-by: Sergio Ahumada <sergio.ahumada@nokia.com>