testrunner: fixed hang/crash info missing for flaky tests sometimes
If a test timed out or exited due to a signal, we were only printing the
info about it if the overall status was a failure.  Therefore, if a test
hung on the first attempt and succeeded on the second attempt, we'd
incorrectly omit the "Timed out" and "exited with signal" messages.

Make sure instead that, if we run the test more than once, we print out
this info after every attempt.

Change-Id: Ifb6d2fe7b0223c1da3dfa42ff897e47ace410f89
Sanity-Review: Qt Sanity Bot <qt_sanity_bot@ovi.com>
Reviewed-by: Jyri Tahtela <jyri.tahtela@nokia.com>
Reviewed-by: Sergio Ahumada <sergio.ahumada@nokia.com>