tst_licenses.pl: allow excluded and optional files
%excludedFiles was renamed to %optionalFiles to reflect that these files
may optionally contain a license header - if they do, they are not
excluded from the check.  The previous name of %excludedFiles was not
consistent with @excludedModules (which are entirely skipped).

A new %excludedFiles was introduced which allows entirely skipping
checks on some files.  Currently the only known use case for this is
to support testing of the tst_licenses.pl script itself.

Change-Id: I87a32c8e900aa63014338cddb637f3c49c37f219
Sanity-Review: Qt Sanity Bot <qt_sanity_bot@ovi.com>
Reviewed-by: Jason McDonald <jason.mcdonald@nokia.com>