Remove "All rights reserved" line from license headers.
As in the past, to avoid rewriting various autotests that contain
line-number information, an extra blank line has been inserted at the
end of the license text to ensure that this commit does not change the
total number of lines in the license header.

Note that the "All rights reserved" line has not been removed from the
test data for the license checker's selftest, but the extra blank line
has been added to those files.

Change-Id: Icf3080dcdb51c2cbf0e825f2f374cb9666eea44c
Sanity-Review: Qt Sanity Bot <qt_sanity_bot@ovi.com>
Reviewed-by: Rohan McGovern <rohan.mcgovern@nokia.com>