Removed python support.
setup.pl and test.pl were originally ported from another repository
which contained a mixture of python and perl scripts.  This repository
only contains perl, but the python functionality was retained so that
writing scripts in python would still be possible.

However, that has not happened, and some are finding the python setup
requirements to be cumbersome.  Remove the feature.

Change-Id: If7f8c2cd128568288aa1d8d506b647dfb37746e9
Sanity-Review: Qt Sanity Bot <qt_sanity_bot@ovi.com>
Reviewed-by: Toby Tomkins <toby.tomkins@nokia.com>