parse_build_log.pl: fixed extremely bad performance in some cases
If the log contained a line beginning with a huge amount of whitespace,
one particular regular expression would have extremely bad performance
(for all intents and purposes, a complete hang).

That regex was written to permit arbitrary amounts of whitespace at the
beginning and end of the line.  Fix that to something more reasonable.
This is probably a good convention to follow when adding new regexes
(avoid using the quantifiers with no upper bound).

A maximum line length was also added.  The log parser will not scan any
lines longer than this.

This bug was triggered by our switch to the clang compiler on mac,
which has output a line with ~4KB of whitespace when compiling
qtdeclarative (log included as testdata).

Change-Id: I9c6170a5a80a8b98354e2fcdf12415786e7b2eec
Sanity-Review: Rohan McGovern <rohan.mcgovern@nokia.com>
Reviewed-by: Toby Tomkins <toby.tomkins@nokia.com>