qtmod_test.pl: don't invoke qmake with -r
We were invoking qmake with -r in the case where the tested module was
hosted outside of qt5.git, but not in other cases.  That is not only
inconsistent, but also often doesn't work - qmake may run syncqt and
syncqt may generate a .qmake.cache file, but that .qmake.cache file
would not take effect in the running `qmake -r', almost certainly
resulting in strange build failures.

Change-Id: Ie95add2f3cdde81157d279e3ade7a6d5c4090ba4
Reviewed-by: Toby Tomkins <toby.tomkins@nokia.com>