CodeCoverage: Report untested source files in html report.
Add a new step in xml2html_testcocoon.pl script.

A new html report will be created including every sources files (.c and .cpp)
found in the include path but not found in the coverage analysis. Files found
under the exclude paths and generated files moc_*.cpp and qrc_*.cpp
will not be part of that report.

This will help identify the files that are not tested in the current
configuration because they are not compiled. For example because a feature is
not enabled or because some files are only compiled on a specific platform.

Add autotests for xml2html_testcocoon.pl.

Task: QTQAINFRA-8

Change-Id: I01c9a89e234607135b5f72d40d30c2e54dd80f7d
Reviewed-by: Rohan McGovern <rohan.mcgovern@nokia.com>