qtmod_test.pl: allow overriding command and args for `make check'
Set make-check.bin and make-check.args to specify the make command, and
arguments, to be used while doing `make check'.

The primary use cases are to use a different level of parallelization
(-j) for tests vs compilation, and to allow switching between nmake and
jom or vice-versa when running the tests on Windows.

Change-Id: I740472a881b674392dddb15fdb6bae4485a9af90
Reviewed-by: Toby Tomkins <toby.tomkins@nokia.com>