qtmod_test.pl: don't replace `jom' with `nmake' for default make check
This seemed like a good idea at the time, but in fact it is unsafe,
because the user may have specified some arguments in make.args which
can be understood by jom but not understood by nmake (in particular,
-j).

Trying to do anything clever here will probably be counter-productive.
Just leave it to the caller to specify the correct make-check values.

Change-Id: I107e2664ee5530221cf0a0f47d51e0afab34ca43
Reviewed-by: Toby Tomkins <toby.tomkins@nokia.com>