Fixed bugs in stdout_cb/stderr_cb callbacks on Windows
The first parameter of the callback is supposed to be the handle (STDOUT
or STDERR), not the Proc::Reliable object.  This was unnoticed because
none of the users of stdout_cb/stderr_cb on Windows are using that
parameter.

The callback would be activated one time more than necessary, with empty
text, due to line splitting regex matching on \z (zero-width end of
string).

Change-Id: Iccadc4ece285162f3cc7a2b764a694484929ceb0
Reviewed-by: Kalle Lehtonen <kalle.ju.lehtonen@nokia.com>
Reviewed-by: Toby Tomkins <toby.tomkins@nokia.com>
Reviewed-by: Jyri Tahtela <jyri.tahtela@nokia.com>
Reviewed-by: Sergio Ahumada <sergio.ahumada@nokia.com>