Implemented proper stdout_cb/stderr_cb callbacks on Windows.
The initial implementation of QtQA::Proc::Reliable::Win32 used a trivial
and just-barely-good-enough implementation for the stdout/stderr callbacks.
That is no longer good enough; we need proper stdout/stderr capturing on
Windows to implement output buffering for parallel tests, and to implement
the testrunner --tee-logs feature.

Therefore, implement stdout_cb/stderr_cb with the same semantics as
on Linux and Mac.

The main caveat of the implementation is the use of interpreter threads;
there is one thread used to run the process, one thread to read STDOUT,
and one thread to read STDERR.  The usage of threads should not matter
in the majority of cases, but some code (e.g. Test::More) may be
affected by the presence of multiple threads.

Change-Id: Ic3e80afd9b61c79a4c5557c40a7323d1339ba9ef
Reviewed-by: Jyri Tahtela <jyri.tahtela@nokia.com>
Reviewed-by: Sergio Ahumada <sergio.ahumada@nokia.com>
Reviewed-by: Toby Tomkins <toby.tomkins@nokia.com>
Reviewed-by: Kalle Lehtonen <kalle.ju.lehtonen@nokia.com>