Fixed QtQA::Proc::Reliable::Win32 when passed output from Getopt::Long
In some cases, Getopt::Long (since 2.37) does not return plain scalars,
but rather returns complex objects which happen to stringify to the
values passed on the command-line.  This is documented.

Normally this is no problem, but QtQA::Proc::Reliable::Win32 needs to
serialize the arguments to send to another process, and coderefs can't
be serialized.  So, stringify everything ourselves.

Change-Id: Id4b200cf27701fedec44ef7ec980028a1c3a2a9f
Reviewed-by: Jyri Tahtela <jyri.tahtela@nokia.com>
Reviewed-by: Sergio Ahumada <sergio.ahumada@nokia.com>
Reviewed-by: Kalle Lehtonen <kalle.ju.lehtonen@nokia.com>
Reviewed-by: Toby Tomkins <toby.tomkins@nokia.com>