testrunner: add --sync-output option for parallel tests
When --sync-output is used and multiple tests are run in parallel, the
testrunner instances will co-operate with each other to ensure the
output from the tests is not interleaved.

As well as a new test added specifically for this feature, --sync-output
was added to a few existing tests to ensure it does not interfere with
existing features.

Change-Id: I24d57c36739edbc75d560ca5fe6e100adbfe74f5
Reviewed-by: Jason McDonald <jason.mcdonald@nokia.com>
Reviewed-by: Kalle Lehtonen <kalle.ju.lehtonen@nokia.com>