setup.pl: make perl module installation safer and customizable
Use `require' instead of `perl -m' to check installed modules.
Certain modules are not intended to be imported directly, and hence may
give a fatal error when imported by `use'; `require' is safer and
faster, as it checks if the module exists (and can be compiled) but
does not import it.

Add %CPAN_MODULE_META allowing for minor customizations of the
installation process where needed.  This is used to work around one
problem with the Inline::C installer on OSX.

Change-Id: Ifeb3f4c9d17b16536ec86ddb98e1f2830ecaa632
Reviewed-by: Kalle Lehtonen <kalle.ju.lehtonen@nokia.com>