testrunner: add crashreporter plugin, for crash logs on mac
Currently we have the ability to obtain backtraces from crashing tests
on Linux and Mac by using core dumps and gdb.  However, in practice,
core dumps on Mac are too slow, forcing us to leave this feature
disabled.

Add a new testrunner plugin which uses the built-in CrashReporter
service on OS X.  This does not give as much detail as running gdb on a
core dump, but it is fast and it does include a backtrace of each
thread.

Change-Id: Ic2c80d8f8bf72653caa09bad776e7036e3efc762
Reviewed-by: Kalle Lehtonen <kalle.ju.lehtonen@nokia.com>