testrunner: remove sync-output optimization for single running test
Previously, if the --sync-output option were used, the first test to
obtain the output lock would directly output with no buffering.
This ensured that, whenever multiple tests were running, at least one
test would output some progress information to give some assurance that
the whole testrunning process hadn't locked up.

However, this had an overall negative effect on throughput.  If the
first test to grab the lock was a very slow test, other tests might
end up blocking for a long time before they can output and exit.

Note that this also means that stdout and stderr from a test are merged
(in the --sync-output case) even when only a single test is running.
Since this significantly affects the behavior of some autotests which
previously used --sync-output but were not directly attempting to test
--sync-output, the option was removed from those tests.

Change-Id: I06d29922fd9fff97c8ec1d6962694b555d984475
Reviewed-by: Kalle Lehtonen <kalle.ju.lehtonen@nokia.com>
Reviewed-by: Toby Tomkins <toby.tomkins@nokia.com>