testrunner: add --verbose switch to make test output identifiable
Previously, when running many tests consecutively and combining the
output in a single log, there has been no foolproof way to determine
where the output of a particular test begins and ends.

We have been able to guess with high accuracy by parsing some output
from make during `make check', but we may soon stop using make to run
the tests, which means this will no longer work.

Fix this by adding a --verbose option to testrunner.  This surrounds the
test's output with a 'begin' and 'end' line so that it can be
unambiguously determined where a test's output begins and ends.

Change-Id: If8b0f6e6a36afeb85d9886271b586308f00cd90c
Reviewed-by: Toby Tomkins <toby.tomkins@nokia.com>