Fixed testplan where `check' target does a chdir
If the `check' target in the Makefile performed a chdir before calling
the $(TESTRUNNER), as is common for debug-and-release on Windows, the
testplanner would not be able to find the Makefile.  Fixed and added
autotest for this case.

These cases would also generally have a TARGET pointing to another
directory (e.g. ../release/tst_mytest1), which is ugly for display
purposes, so we introduce a new `label' attribute to the testplan.

The testscheduler is modified to prefix all keys with '_', to avoid
clashing with any further metadata added to the testplan in the future.

Change-Id: Ibe9077a90e531324a0ce57a7e9b400787f3edfa8
Reviewed-by: Toby Tomkins <toby.tomkins@nokia.com>
Reviewed-by: Kalle Lehtonen <kalle.ju.lehtonen@nokia.com>