qtmod_test.pl: fixed logic error preventing qtqa tests over submodules
The commit which introduced Qt4 support also introduced this test:

  if (($qt_gitmodule ne 'qt5') or ($qt_gitmodule ne 'qt')) { ... }

...which is always true ($qt_gitmodule cannot be equal to both 'qt5' and
'qt').  This means the else {} path, which was intended to test all qt5
submodules, was dead code.

Change it to eliminate the test altogether: we will simply _always_
iterate over the submodules (and expect that there usually are none).

Task-number: QTQAINFRA-213
Change-Id: I44d86e7c9f3a0e3b6dbc00d348dbc887175ad615
Reviewed-by: Toby Tomkins <toby.tomkins@nokia.com>