qtmod_test.pl: accept an existing qt.dir
Instead of giving a fatal error if pointed at an existing qt5 directory,
just re-use it.  This can be used to avoid recloning qt5 each time.

However, note that using qtmod_test.pl in this way means that the caller
is responsible for ensuring the qt5 repository is cleaned appropriately
before the test begins.

Change-Id: Ic49cbca84468e0b7fa4db91cab80d082d528ecdf
Reviewed-by: Toby Tomkins <toby.tomkins@nokia.com>