Added "git-qtqa-combine" for maintaining vendor branches
This script can take the sources from public git repositories and
combine them with sources from other repositories, retaining the content
and the git history from all repositories.

It is intended to be used for vendor branches of Qt, consisting of the
upstream source code merged with a few device-specific files on top
(e.g. packaging scripts).

Change-Id: I9a1d15c67082e02f7cdee702e39df78fc2c905d9
Reviewed-by: Sergio Ahumada <sergio.ahumada@nokia.com>