testplanner: really fix the "ambiguous makefile on windows" problem.
When attempting to find the calling makefile during "make check" on
Windows, the previous heuristic used was to assume that the makefile
with the shortest name was the "primary" makefile for the test.

That gives incorrect results for tests which used .pro files as subdirs
to bundle some helper projects and the main test project under a single
directory:

  - mytest.pro
    -> mytesthelper1
    -> mytesthelper2
    -> tst_mytest.pro

In the above example, the correct makefile for "make check" would be
named Makefile.tst_mytest.

Give up on hueristics-based approaches.  Instead, some Windows-specific
code is introduced to query the command-line of the parent process(es)
in order to determine the correct makefile.

Change-Id: I138659d1f5585d46ea9032050b550402d93077d2
Reviewed-by: Kalle Lehtonen <kalle.ju.lehtonen@nokia.com>