testplanner: avoid duplicate test labels
There are a handful of tests within Qt5 with duplicate names (e.g.
tst_examples in qtdeclarative and qtquick1).  So far, in the test
summary, these were indistinguishable from each other.

Fix it; use the test directory and arguments to append something
meaningful to disambiguate the tests.  For the above example, this
results in:

  tst_examples (qtdeclarative)
  tst_examples (qtquick1)

Change-Id: Ie866e660972444f611d58ef8d3e12f80fcf7f141
Reviewed-by: Kalle Lehtonen <kalle.ju.lehtonen@nokia.com>