testscheduler: added parallel-stress mode
Running testscheduler in parallel-stress mode will run all tests
concurrently and attempt to give some hint whether or not the tests can
be considered parallel-safe.  A test which fails if run concurrently and
passes if run serially is considered parallel-unsafe; a test which
passes when run concurrently is considered parallel-safe.

The tests are run concurrently against themselves.  For example, if run
with -j5 on qtbase, five parallel instances of tst_qfile will be
spawned; if (and only if) they all pass, tst_qfile is considered
parallel-safe.

The idea behind parallel-stress mode is that running a test concurrently
with itself is the simplest way to determine if a test expects to have
exclusive access to some resource.  However, the output is only an
estimate.

Change-Id: I2dbe6da87e0ff7da21a7ebe0082e8bce067dbce6
Reviewed-by: Toby Tomkins <toby.tomkins@nokia.com>
Reviewed-by: Kalle Lehtonen <kalle.ju.lehtonen@nokia.com>