qt_update_submodules: check for already used Change-Id
The logic is to attempt to re-use the SHA1 of the bot's last commit as
the Change-Id of the next commit.  This breaks if someone else has
already done this - most likely by "stealing" one of the bot's changes
in gerrit.

If the Change-Id we want to use is already used, fall back to a random
one.  This situation should be avoided, since it can result in the bot
having multiple open changes in gerrit at the same time.

Change-Id: Ibc7f2359e63e4a9cb196e15003be846a9e6ce796
Reviewed-by: Toby Tomkins <toby.tomkins@nokia.com>
Reviewed-by: Kalle Lehtonen <kalle.ju.lehtonen@nokia.com>