testrunner: fixed race on --capture-logs file creation
If multiple tests with the same name were run in quick succession, it
was possible for multiple instances of the testrunner to attempt to use
the same output log file.

On Windows, this would cause a fatal error in the test who lost the
race.  On other platforms, it would cause output to be lost from the
test who won the race.

A test was added specifically for this race condition.

Change-Id: I197440a726acb58d615c03b454aef8d6f2aee777
Reviewed-by: Kalle Lehtonen <kalle.ju.lehtonen@nokia.com>