Fixed ambiguous handling of YAML errors in test logs.
We were using the standard "..." marker to denote the end of a YAML
document.  Unfortunately, a few other places would also generate a line
containing "..." and nothing else, making it hard to parse unambiguously
(at least when parsing lines in reverse order).

Add a comment as well, which is permitted by the YAML spec and makes the
formatting unambiguous.

Change-Id: Ie0f3d26465f65c4ccf664a81cd05f38f29f79a9c
Reviewed-by: Kalle Lehtonen <kalle.ju.lehtonen@nokia.com>