Make testrunner/testscheduler output unambiguous.
Make testrunner's begin/end lines include the same test label as used by
testscheduler, so that the output can be reliably matched up.

Make parse_build_log use this test label when identifying a test's
output, so that we won't find the wrong test's output.  Previously, we
assumed that the test's output is always immediately adjacent to the
failure message from testscheduler, which is sometimes false for
parallel tests.  Includes some refactoring to parse_build_log's chunk
handling.

Note that two testdata copies were added for this issue, one of them
manually edited to make the output of tst_qchar shorter (that test
outputs a very large amount of text, which gets in the way of testing
this bug), and both of them edited to include the new style of
testrunner output.

Task-number: QTQAINFRA-534
Change-Id: I7aa23235819c4c2c028adb1e1c416459b93e8289
Reviewed-by: Kalle Lehtonen <kalle.ju.lehtonen@nokia.com>