testscheduler: limit size of the "run concurrently with" output
When a test fails, we want to say which other tests were running
concurrently.  However, if a test runtime is high, this information can
be very large and even exceed the parse_build_log per-line limit.

Limit the amount of tests which will be listed in this message.
Tests are omitted from the middle of the message, since it is generally
possible to extract those tests from the log if the first and last run
tests are known.

Change-Id: Ie0792ca7a65aca142d3a928073f067c6b0dcb45f
Reviewed-by: Kalle Lehtonen <kalle.ju.lehtonen@nokia.com>