New Jenkins<->Gerrit integrator.
This script implements a quality-gate workflow, using Gerrit as the
source of commits to be tested, and Jenkins as the testing tool.

Features / improvements over the earlier integrator include:

  - a single process manages all projects, rather than one process per
    project

  - no polling required (where network architecture permits); uses
    gerrit stream-events and jenkins post-build events

  - no git checkouts required on the machine running the integrator

  - warnings and errors are automatically buffered and sent as emails

  - a read-only remote API is available to retrieve the CI state in JSON
    format

This replaces the qt-jenkins-ci.pl script called from within Jenkins,
inverting the responsibility; Jenkins is contained within the CI
workflow rather than the CI workflow being contained within Jenkins.

Since the new integrator requires a lot of functionality in common with
other Jenkins-related scripts, some of this was extracted out into new
QtQA modules.

Change-Id: I8bf28ba9885b04935181b0c48ae2c35c211cce91
Reviewed-by: Sergio Ahumada <sergio.ahumada@digia.com>