parse_build_log: replace File::Fetch with AnyEvent::HTTP
File::Fetch doesn't work with https, doesn't have a configurable
timeout, and has a few odd quirks.

The lack of a timeout setting is the main motivation for this change, as
parse_build_log.pl currently will hang for a very long time if run
against e.g. an aggressively firewalled server.

Change-Id: I7c978d07e7776d9c8796c7314dc0d39401463d58
Reviewed-by: Kalle Lehtonen <kalle.ju.lehtonen@nokia.com>