QtQA::TestScript: ensure $? is set when exe(), exe_qx() return
Set $? to the right value after exe() and exe_qx(), for compatibility
with system() and qx().

Actually, $? would often end up being set to the correct value already,
but we did not enforce, test or document it, and the usage of
QtQA::Proc::Reliable could leave the value set incorrectly.

Change-Id: Idd0710ec8afb63ae55a9a0e9d0bbeef40b64f5e6
Reviewed-by: Toby Tomkins <toby.tomkins@nokia.com>
Reviewed-by: Kalle Lehtonen <kalle.ju.lehtonen@nokia.com>