git-qtqa-combine: fixed for newer versions of git
At some point between git 1.7.0.4 and 1.7.9.5, the behavior of
"git checkout <tree-ish> -- file1 file2" changed incompatibly. In the
earlier version, if any of the named files did not exist in <tree-ish>
but _did_ exist in HEAD, the command would succeed and behave as though
the missing files were not present on the command line. In the newer
version, it is now a fatal error if any named file is not present in
<tree-ish>.

Therefore, when we are using 'git ls-files' to list the files from
a source ref we are going to checkout, use the --modified flag to only
list those files which are modified between the source ref and HEAD.

Fixes 10-git-qtqa-combine.t autotest on Ubuntu 12.04.

Change-Id: I8b25237c31bd65c8ba58b1e979fd8feea916ef65
Reviewed-by: Kalle Lehtonen <kalle.ju.lehtonen@nokia.com>
Reviewed-by: Sergio Ahumada <sergio.ahumada@nokia.com>