qt-jenkins-ci: fixed update of latest, latest-success links
We were running ln with the `-f' parameter with the intent of
overwriting existing links. However, if the link target was a directory,
ln would dereference the link and create a link inside the directory.

Therefore, for a new project, we'd create the first link:

  latest -> build_00001

... as expected, but subsequent links would be created _inside_ of the
'latest' (actually build_00001) directory:

  build_00001/build_00002 -> build_00002
  build_00001/build_00003 -> build_00003
  ...

This is fixed by passing the -n / --no-dereference option to the ln
commands.

Change-Id: I1f188567d2580451236d5b90f86a737143860bb7
Reviewed-by: Kalle Lehtonen <kalle.ju.lehtonen@nokia.com>