parse_build_log: added YAML output option
Formatting the output as a YAML document makes it possible for a calling
script to unambiguously extract metadata from the log.  Currently, the
YAML document only contains the summary and detail, with additional
metadata to be added in later commits.

A selection of plain text testdata was copied and transformed into YAML
testdata.

Change-Id: Ia9a741e7682d439e4a99d9508857b4a0a252eddf
Reviewed-by: Sergio Ahumada <sergio.ahumada@nokia.com>