parse_build_log: added should_retry metadata to YAML output
parse_build_log knows about various error cases which are unlikely to be
caused by the code under test. In plain text mode, a hint is written
that it might make sense to retry the build. Add the same hint to the
YAML mode. A calling script can use this to repeat a failed build.

Change-Id: I55e9f620f7b365e09062ba550527e4d85bd456ac
Reviewed-by: Sergio Ahumada <sergio.ahumada@nokia.com>