summarize-jenkins-build: allow treating abortions as failures
If a human aborted a build in progress, it makes sense to simply report
the build as "ABORTED" (perhaps along with the name of the aborter).

If the CI system aborted a build because it saw that some build
configurations had already failed, it makes sense to ignore the
"ABORTED" status and report the failing configurations instead.

This commit adds a command-line option, --ignore-aborted, to toggle
between the two above behaviors.

Change-Id: I31fc984ea64ac586f0dd4d004a8602e5c14354d1
Reviewed-by: Sergio Ahumada <sergio.ahumada@nokia.com>