summarize-jenkins-build: support YAML output format
Similar to the recent change introducing YAML support to
parse_build_log.pl, this change allows YAML output so that the caller
can unambiguously parse the output of this script.

Metadata from parse_build_log.pl is passed through unaltered; most
notably, whether or not the build should be retried.

Change-Id: I77f928e4dacc77117244fa551d0ba1956fb110cd
Reviewed-by: Sergio Ahumada <sergio.ahumada@nokia.com>