parse_build_log: handle failure-like messages in test output
It might be that a testcase which passes generates some output matching
some of the failure patterns. For example, qtbase has a 'cmake' autotest
which invokes the compiler and may generate compiler errors.

Prior to this commit, such errors from within a test would be extracted
even if the test passed. This could result in an incorrect failure
summary.

Change to never scan within a test's output for other kinds of errors,
whether the test passes or fails.

Change-Id: I7d4845860cf02319a82177dd9a3f1eacff769fe5
Reviewed-by: Sergio Ahumada <sergio.ahumada@digia.com>